python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
segment-anything>=1.0.0
diffusers>=0.20.0
transformers>=4.20.0
requests>=2.25.0
tqdm>=4.62.0
matplotlib>=3.6.0
scikit-image>=0.20.0
gdown>=4.7.0
omegaconf>=2.3.0
